requirements.txt
alembic==1.7.4
astroid==2.8.4
autopep8==1.6.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.8.0
pylint==2.11.1
python-dotenv==0.19.1
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
typing-extensions==3.10.0.2
Werkzeug==2.0.2
wrapt==1.13.3
